feat: Added Japanese translation for the application and Factorio 1.1.x (#1403)
Nice to meet you. I am a fan of Factorio and a Japanese translator
(Zefard) of the official Wiki.
Since FactorioLab did not yet support Japanese, I created Japanese
translation data for 1.1.

Since I am also an accomplished Angular programmer, it was easy to get
FactorioLab up and running on my PC. It appears to be working well in my
browser.

If you like, could you merge this translation with the official one?


![image](https://github.com/factoriolab/factoriolab/assets/26034872/59663ced-1ee3-430e-b5b9-10bf28c577a0)

![image](https://github.com/factoriolab/factoriolab/assets/26034872/371a1d53-6981-4d5c-891d-c03314e560a9)