fix: add prod bonus from research to quality recipes, closes #1635 (#1653)
By generating the list `canProdUpgradeRecipeIds` after the quality
recipes are generated and not before, I can now also set the
productivity bonuses for the quality variants of the recipes with
researchable prod bonus. This fixes #1635.

However it also introduces a new recipe prod bonus setting for each
quality variant of the base recipe (now 86 settings vs 18 now and only 8
researches). But this would be resolved when fixing #1643, and for me it
works great even in the current state (finally I can calculate LDS and
blue circuit recycling loops).