build: Adding Translation helper script  (#1671)
The translation helper script compares a source (usually en.json) and a
target JSON, finds the keys that are already translated and those that
are missing, and then creates a temporary file containing the missing
keys set to UNTRANSLATED.

It then prompts the user to translate the missing keys. When there are
no more untranslated keys in the temporary JSON, that file is renamed to
replace the target JSON.

The script can be gracefully exited after every 10 keys and will save
any incomplete translations in the temporary file. On subsequent starts,
if a temporary file is present, it will skip the comparison with the
source.

This is the simplest way I have found to help with translation, but I
would gladly make changes to the script or even to its workflow.