refactor: Extract mod set "defaults" as separate files (#1676)
At least for Factorio, most of `src/data/*/data.json` is automatically
generated from game data. However it also contains a `defaults` key that
needs to be authored and maintained manually. This can be tricky since
`data.json` is one huge single line without whitespace.

With this PR, the source of truth for defaults is instead in new
`src/data/*/defaults.json` files (initially extracted with `jq .defaults
< $modset/data.json > $modset/defaults.json`) that can use multiple
lines and indentation. This will make future changes easier to review.
I’d like to change the defaults and presets for Space Age in another PR.

Defaults are still copied into `data.json` so that the frontend does not
need to make an extra HTTP request. This copy happens both as part of
the `factorio-build` script, and in a new `update-defaults` script. This
new script does not need `factorio-dump` to have been run just before,
and works for all games not just Factorio.

Wikis don’t have pull requests, so
https://github.com/factoriolab/factoriolab/wiki/Creating-a-new-factorio-mod-set
should also be updated to add a step at the end of the "Modifying
FactorioLab" section:
> 4. Create a `src\data\abc` directory and `src\data\abc\defaults.json`
file, replacing `abc` with your new mod id. Use
`src\data\2.0\defaults.json` (or the most similar existing modset) as a
starting point and tweak it as needed.