feat: Rewrite presets for 2.0 and Space Age
The "Factory" section of settings has a few "presets" that change
defaults for the rest of that section. For Factorio, these presets
are well suited for the 1.1 base game but less so for Space Age
or even 2.0 base game.

This PR replaces presets and **changes the meaning of some existing URLs**.

New presets for the 2.0 base game:

* Minimum: (yellow) transport belt, stone furnace, assembling machine 1,
  no modules or beacons
* Upgraded: (blue) express belt, electric furnace, assembling machine 3,
  no modules or beacons.
* 2 beacons: prod 3 modules in machines, speed 3 in beacons
* 8 beacons
* 12 beacons

Opinionated changes:

* The minimum furnace is stone, not electric
* There is no longer a "modules" preset that defaults to prod modules
  (and their speed penalty) without any (speed modules in) beacons
  to compensate.
* Using few beacons is much better than in 1.1.
  8 or 12 beacons presets are preserved for those
  who want to maximize speed despite diminishing returns.
  4 beacons (one-sided solid row) might be another compromise.

New presets for Space Age:

* Minimum: (yellow) transport belt, stone furnace, assembling machine 1,
  no modules or beacons
* Upgraded: (blue) express belt, electric furnace, assembling machine 3,
  no modules or beacons.
* 2 beacons: (green) turbo belt, all new machines preferred
  (except biochamber, listed after chem plants),
  prod 3 modules in machines, speed 3 in beacons
* Legendary: same as 2 beacons but everything in legendary quality

Opinionated change: this time high-beacon presets are removed
because they make even less sense.
As quality increases it doesn’t take many beacons for inserters
to become a bottleneck for moving items in and out of machines fast enough.
(Of course it’s still possible to manually configure more beacons.)

Tweak suggestions welcome.

----

Presets for Factorio 1.1 or for other games are unchanged.

Internally, a new mechanism allows a mod set to define custom presets
in `defaults.json` instead of the existing ones defined in TypeScript code.